requirements.txt
slack-bolt>=1.21,<2
slack-sdk>=3.33.1,<4
# If you use a different LLM vendor, replace this dependency
openai

pytest
flake8
black
python-dotenv
